change the way of getting studio version
I changed the way of getting version of studio because there are some problems for the previous commit.

It doesn't work because the org.talend.commons.utils.VersionUtils class is in fact in code generating (eclipse) level. So we will not have it in the generated project in Studio, so I moved it (version retrieving) to javajet level as shown in the 3 other related commits, to get the version correctly.